// src/lib/types/api.ts
export interface Researcher {
  id: number
  name: string
  email: string
  grade: string
  teamId?: number
  role: "admin" | "researcher"
}